Updated readme given that "mid price" has been changed to "last price"
I also removed "quartiles" as one of the key goals of the package. I believe PDF and CDF are the key goals of the project, and quartile is only a description of the PDF/CDF.

e.g. we could also generate descriptions of the variance, skew, kurtosis of the stock, and I think it would be accurate but unnecessary to say the goal of the package is to "generate the PDF, CDF, quartiles, variance, skew, kurtosis, ..."